refactor(core): Move structured-output error helper to a util file
Addresses review feedback: extract describeStructuredOutputError out of agents.service.ts into utils/structured-output-error.ts, with a dedicated unit test. Behaviour is unchanged.
